Move Auth controllers out of subdirectory
