+++ routes/web.php
@@ -8,8 +8,8 @@ use App\Http\Controllers\RequestRecordsController;
 use App\Http\Controllers\DashboardController;
 use App\Http\Controllers\FeedbackController;
 use App\Http\Controllers\VulnerabilityTogglesController;
-use App\Http\Controllers\Auth\AuthenticatedSessionController;
-use App\Http\Controllers\Auth\RegisteredUserController;
+use App\Http\Controllers\AuthenticatedSessionController;
+use App\Http\Controllers\RegisteredUserController;
 use App\Http\Controllers\PatientInfoController;
 
 Route::middleware('guest')->group(
